requirements.txt
# Streamlit 대시보드
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
numpy>=1.24.0

# FastAPI 서버
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# LangChain (실제 사용 패키지만)
langchain==0.3.21
langchain-core==0.3.46
langchain-openai==0.3.9
langchain-text-splitters==0.3.7
langchain-upstage>=0.1.0
langchainhub==0.1.21

# Vector DB
langchain-pinecone>=0.1.0
pinecone-client>=3.0.0
tiktoken>=0.5.0
openai>=1.0.0

# 데이터 시각화 & 분석
matplotlib>=3.6.0
seaborn>=0.12.0
openpyxl>=3.1.0

# HTTP 요청
requests>=2.28.0

# 환경 변수 & 이미지
python-dotenv>=0.19.0
Pillow>=9.5.0

# PDF/DOCX 처리
PyMuPDF>=1.23.0
pdfplumber==0.11.1
python-docx==1.1.2
reportlab
